refactor anthropic helpers into dedicated modules
